+++ .storybook/components/icons/FlashIcon.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+export const FlashIcon = () => (
+    <svg
+        className='sidebar-icon'
+        width='24'
+        height='24'
+        viewBox='0 0 24 24'
+        role='img'
+        focusable='false'
+        fill='currentColor'
+    >
+        <path d='M7.5 12L14.5 5L12.9444 10.4444L16.5 12L9.5 19L11.3889 13.5556L7.5 12Z' />
+    </svg>
+);