+++ packages/calendar-input/src/index.ts
@@ -1,2 +1,4 @@
-export * from './Component';
+export { CalendarInputResponsive as CalendarInput } from './responsive';
+export type { CalendarInputResponsiveProps as CalendarInputProps } from './responsive';
+
 export * from './utils';