+++ packages/calendar-range/src/components/divider/Component.tsx
@@ -1,0 +1,23 @@
+import React, { FC } from 'react';
+import cn from 'classnames';
+
+import { CalendarRangeProps } from '../../Component';
+
+import styles from './index.module.css';
+
+type Props = {
+    inputFromProps?: CalendarRangeProps['inputFromProps'];
+    inputToProps?: CalendarRangeProps['inputToProps'];
+};
+
+export const Divider: FC<Props> = ({ inputFromProps, inputToProps }) => {
+    const outer =
+        inputFromProps?.label &&
+        inputFromProps?.labelView === 'outer' &&
+        inputToProps?.label &&
+        inputToProps?.labelView === 'outer';
+
+    const size = inputFromProps?.size || inputToProps?.size || 's';
+
+    return <span className={cn(styles.component, styles[size], { [styles.outer]: outer })} />;
+};