+++ packages/date-range-input/src/index.ts
@@ -1,1 +1,2 @@
-export * from './Component';
+export { DateRangeInputResponsive as DateRangeInput } from './responsive';
+export type { DateRangeInputResponsiveProps as DateRangeInputProps } from './responsive';