+++ packages/date-time-input/src/index.ts
@@ -1,1 +1,2 @@
-export * from './Component';
+export { DateTimeInputResponsive as DateTimeInput } from './responsive';
+export type { DateTimeInputResponsiveProps as DateTimeInputProps } from './responsive';