+++ packages/input-autocomplete/src/index.ts
@@ -1,1 +1,3 @@
-export * from './Component';
+export { InputAutocompleteDesktop as InputAutocomplete } from './Component.desktop';
+export type { InputAutocompleteDesktopProps as InputAutocompleteProps } from './Component.desktop';
+export * from './responsive';