+++ packages/modal/src/Component.desktop.tsx
@@ -1,5 +1,8 @@
 import React, { forwardRef } from 'react';
 
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import type { BaseModalProps } from '@alfalab/core-components-base-modal';
+
 import { Closer } from './components/closer/Component';
 import { Content } from './components/content/Component';
 import { Footer } from './components/footer/Component';