+++ packages/picker-button/src/index.ts
@@ -1,2 +1,3 @@
-export * from './Component.responsive';
+export { PickerButtonResponsive as PickerButton } from './responsive';
+export type { PickerButtonResponsiveProps as PickerButtonProps } from './responsive';
 export type { PickerButtonDesktopProps, PickerButtonSize, PickerButtonVariant } from './Component';