+++ packages/chart/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 3.1.5
+
+### Patch Changes
+
+### [#418](https://github.com/core-ds/core-components/pull/418)
+
+-   Исправлена проблема с default-импортом в cjs форматах
+
+-   Обновлены зависимости
+    -   typography@3.0.7
+
 ## 3.1.4
 
 ### Patch Changes