+++ packages/tabs/CHANGELOG.md
@@ -1,5 +1,41 @@
 # Change Log
 
+## 7.2.8
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   picker-button@10.2.1
+
+## 7.2.7
+
+### Patch Changes
+
+### [#418](https://github.com/core-ds/core-components/pull/418)
+
+-   Исправлена проблема с default-импортом в cjs форматах
+
+-   Обновлены зависимости
+    -   picker-button@10.2.0
+    -   badge@4.0.4
+    -   keyboard-focusable@4.0.3
+    -   tag@5.0.3
+
+## 7.2.6
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   picker-button@10.1.1
+
+## 7.2.5
+
+### Patch Changes
+
+### [#388](https://github.com/core-ds/core-components/pull/388)
+
+-   Исправлен импорт PickerButtonDesktop компонента
+
 ## 7.2.4
 
 ### Patch Changes