+++ packages/time-input/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @alfalab/core-components-time-input
 
+## 2.1.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@11.1.2
+
+## 2.1.1
+
+### Patch Changes
+
+### [#418](https://github.com/core-ds/core-components/pull/418)
+
+-   Исправлена проблема с default-импортом в cjs форматах
+
+-   Обновлены зависимости
+    -   icon-button@6.0.2
+    -   input@11.1.1
+
 ## 2.1.0
 
 ### Minor Changes